+++ fjs/cli/proof.f.ts
@@ -1,6 +1,6 @@
 import { pure } from '../effects/module.f.mjs'
 import type { NodeOp, NodeProgramOptions } from '../effects/node/types.ts'
-import { defaultNodeProgramOptions, emptyState, virtual } from '../effects/node/virtual/module.f.ts'
+import { defaultNodeProgramOptions, emptyState, virtual } from '../effects/node/virtual/module.f.mjs'
 import { dispatch } from './module.f.mjs'
 import type { Commands } from './types.ts'
 import { assert, assertEq } from '../asserts/module.f.mjs'